Improve UX/UI of Completed Progress
