+++ src/components/common/DailyHabitTracker.tsx
@@ -4,6 +4,9 @@ import { Habit } from "@/types/habit";
 import DailyHabitDate from "@/components/common/DailyHabitDate";
 import DailyHabitProgressIndicator from "@/components/common/DailyHabitProgressIndicator";
 import DailyHabitList from "@/components/common/DailyHabitList";
+import toggleOnSound from "@/assets/audio/habit-toggled-on.mp3";
+import toggleOffSound from "@/assets/audio/habit-toggled-off.mp3";
+import completedAllHabits from "@/assets/audio/completed-all-habits.mp3";
 
 export default function DailyHabitTracker({
   isDarkMode,
@@ -21,7 +24,7 @@ export default function DailyHabitTracker({
     return Math.round((completedHabits / habits.length) * 100);
   }, [habits]);
 
-  const toggleHabit = (id: string) => {
+  const toggleHabit = async (id: string) => {
     setHabits(
       habits.map((habit) => {
         if (habit.id === id) {
@@ -30,11 +33,27 @@ export default function DailyHabitTracker({
             setAnimatingHabitId(id);
             setTimeout(() => setAnimatingHabitId(null), 1000);
           }
+
           return { ...habit, completed: newCompleted };
         }
         return habit;
       })
     );
+
+    const habit = habits.find((habit) => habit.id === id);
+    const allHabitsCompleted =
+      habits.filter((habit) => habit.completed).length === habits.length - 1 &&
+      !habit?.completed;
+
+    if (allHabitsCompleted) {
+      const audio = new Audio(completedAllHabits);
+      await audio.play();
+    } else {
+      const audio = habit?.completed
+        ? new Audio(toggleOffSound)
+        : new Audio(toggleOnSound);
+      await audio.play();
+    }
   };
 
   return (
